разнес сервера на файлы
