+++ internal/agent/sender_grpc.go
@@ -1,0 +1,109 @@
+package agent
+
+import (
+	"context"
+	"sync"
+	"time"
+
+	"github.com/ncyellow/devops/internal/agent/config"
+	pb "github.com/ncyellow/devops/internal/grpc/proto"
+	"github.com/ncyellow/devops/internal/repository"
+	"github.com/rs/zerolog/log"
+	"google.golang.org/grpc"
+)
+
+type GRPCSender struct {
+	conf   *config.Config
+	conn   *grpc.ClientConn
+	client pb.MetricsClient
+}
+
+// SendMetricsBatch отправляет все метрики одной пачкой на указанный url
+func (g *GRPCSender) SendMetricsBatch(dataSource []repository.Metrics) {
+	// Если метрик данных нет сразу на выход
+	if len(dataSource) == 0 {
+		return
+	}
+
+	var counters []*pb.CounterMetric
+	var gauges []*pb.GaugeMetric
+	for _, metric := range dataSource {
+		// делаем так потому что hash опциональное поле и мы будем передавать указатель и потому отдельная переменная
+		var hash string
+		if metric.Hash != "" {
+			hash = metric.Hash
+		}
+
+		switch metric.MType {
+		case repository.Counter:
+			counters = append(counters, &pb.CounterMetric{
+				Name:  metric.ID,
+				Value: *metric.Delta,
+				Hash:  &hash,
+			})
+		case repository.Gauge:
+			gauges = append(gauges, &pb.GaugeMetric{
+				Name:  metric.ID,
+				Value: *metric.Value,
+				Hash:  &hash,
+			})
+		}
+	}
+
+	resp, err := g.client.AddMetric(context.Background(), &pb.AddMetricRequest{
+		Counters: counters,
+		Gauges:   gauges,
+	})
+	if err != nil {
+		log.Info().Msgf("%s", err.Error())
+	}
+	if resp.Error != "" {
+		log.Info().Msg(resp.Error)
+	}
+}
+
+// SendMetrics отправляет все метрики одной пачкой на указанный url
+func (g *GRPCSender) SendMetrics(dataSource []repository.Metrics) {
+	// Если метрик данных нет сразу на выход
+	if len(dataSource) == 0 {
+		return
+	}
+}
+
+func (g *GRPCSender) Close() {
+	// Общая функция очистки ресурсов. для http не требуется
+	g.conn.Close()
+}
+
+// RunSender запускает цикл по обработке таймера отправки метрик из канала out на сервер
+func RunSender(ctx context.Context, conf *config.Config, out <-chan []repository.Metrics, wg *sync.WaitGroup) {
+
+	repo := repository.NewRepository(conf.GeneralCfg())
+
+	sender := CreateSender(conf)
+	defer sender.Close()
+
+	tickerReport := time.NewTicker(conf.ReportInterval.Duration)
+	defer tickerReport.Stop()
+
+	for {
+		select {
+		case <-tickerReport.C:
+			// Две отправки для совместимости со старой версией, по старому протоколу
+			sender.SendMetrics(repo.ToMetrics())
+			// По новой через Batch
+			sender.SendMetricsBatch(repo.ToMetrics())
+		case metrics, ok := <-out:
+			if !ok {
+				wg.Done()
+				return
+			}
+			for _, metric := range metrics {
+				repo.UpdateMetric(metric)
+			}
+		case <-ctx.Done():
+			wg.Done()
+			return
+		}
+	}
+}
